path-based graph reduction, cluster distance, dendrogram Newick output
- I implemented the new graph creation algorithm (details and questions below)
- I implemented the distance between any 2 clusters, any 2 patients, or a patient and a cluster (command is "distance <id> <id>")
- Implemented outputting the Newick tree format for the main dendrogram as well as for any visualized clusters
- Only display bootstrapping values (output and dendrogram) if bootstrapping is actually performed
- Fixed some output bugs

New Graph Creation Algorithm:
When reducing a graph, I now keep track of what is considered to be valid "endpoints" of paths: i.e. in our case, we want the endpoints to be strictly those genes that are originally input for each patient.
Reduce the set of valid endpoints to include only the ones that are present in the graph
Sort the result of (1) according to some reduction metric. Call this sorted list V.
Currently, that reduction metric is simply the degree of the node
It would be possible (and easy to change this) to use the number of patients the particular endpoint gene was found in
For each gene G in V:
Find all new paths that would be possible if G were to be included in the graph (i.e. all paths between genes with a higher reduction metric than G and G itself)
Count the number of NEW nodes that would be included if ALL paths were to be added to the graph
If the current number of nodes + the number of new nodes <= max nodes: add all paths
Otherwise, stop the algorithm. Note that this means adding G is all or nothing with respect to the paths it will include in the final graph. Also note that the algorithm will NOT try to go to the next gene in V (with a lower reduction metric) and see if it can add the set of paths including the next gene; the algorithm simply stops (which I believe to be the intuitive behavior).
(Random note-to-self: I should deterministically handle tie-breakers for sorting the reduction metric.)